Refactor: Isolate Flickbot and Triggerbot logic into stuff.cpp/h
- Completely removed Flickbot and Triggerbot variables and logic from apex_dma.cpp.
- Reverted AimbotLoop in apex_dma.cpp to the previous normal version as requested.
- Created stuff.h and stuff.cpp to house Flickbot and Triggerbot logic.
- Moved helper functions isAllowedWeapon and IsInCrossHair to stuff.cpp.
- Added FlickbotTriggerbotLoop and UpdateStuffVars to stuff.cpp for isolated execution and synchronization.
- Updated main in apex_dma.cpp to manage the new Flickbot/Triggerbot thread.
- Added #pragma once to all headers in apex_dma/ to prevent redefinition errors.
- Updated Makefile to include stuff.o in the build process.
