+++ apex_dma/Game.h
@@ -1,3 +1,4 @@
+#pragma once
 #include "Math.h"
 #include "offsets.h"
 #include "memory.h"
