+++ apex_dma/Math.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <math.h>
 #include "vector.h"
