+++ apex_dma/prediction.h
@@ -1,3 +1,4 @@
+#pragma once
 #include <cmath>
 #include "Game.h"
